Implement spiral drawing
At first I just picked a color from the center and it just didn't look
good enough on the fine details..

Now it's an actually fancy scanning of all pixels in a slightly larger
radius then the brush.

The algorithm does a Math.pow(value,
2.2ish) when reading the color values and the inverse after the average has
been calculated.

This does a far better job at aproximating intensities in the non-linear
RGB color space.

In the end, it's all fudged of course to better represent the values in
the current image (as well as picking a good spiral center for it).